fix a subtle bug in rendering
while everything looks fine when running a server locally, sometimes classes
would be misapplied in the shadow DOM when deploying to production. my theory
is that React or webpack was mangling the tree struture when <p>s were used
around <div>s; this commit eliminates that particular structure in the DOM and
seems to have fixed the issue.

also add ;s to the end of export in all components for consistency sake