https://github.com/davidbettis/daily-office/issues/2
Fix a bug where the links to the ESV were stale when switching between days. Isolated the issue to the page generation component in Gatsby, but the project seemed a bit complex to fully debug. Instead, migrate to react-static https://github.com/react-static/react-static, which is a far simpler framework (pure React) and a better fit for the needs of this project.